[bringup] Moisture sensor attached to adc pin 0.

+++ src/service/dashboard_pub.h
@@ -91,7 +91,7 @@ namespace service
      *
      * @return double The moisture level in percentage.
      */
-    double convertVoltageToMoistureLevel(double voltage);
+    int32_t convertVoltageToMoistureLevel(int32_t voltage);
 };
